Restructure of the code and examples
