+++ python/phevaluator/evaluator/hashtable6.py
